return error messages when devices are not affected by assignments
The create assignment api would simply return a list of deviceIds that
were affected by the assignment, but if some device was not affected,
for whatever reason (update not compatible, assignment exists, etc)
its deviceId was omitted silently in the response.

These changes make the create assignment api endpoint return an object
describing which updates are affected by the update and which are not,
along with a an error structure describing why an ecu of a device was
not updated.