Allow setting an expiresAt when creating lockbox
Support POST on `offline-uptats/:role` now accepts `expiresAt` in the
json body and sets the lockbox expiresAt with that value. 365 months
from now continues to be the default.

If the `expiresAt` requested by the user is earlier than the latest expire
date of all lockboxes, the user requested `expiresAt` will be ignored
and the latest `expiresAt` of all lockboxes will be used.