Allow ECU replacement when device has running assignments
IIF the new ECU set contains all ECUs for which there is an
assignment.

This is required to deal with with updates that cause devices to
replace their ECU set, adding new ECUS, for example. After such update
is installed, the device tries to register the new ECU set **before**
the new manifest is reported to the server, which means the assignment
is still running. We need a way to allow the device to update the ECU
set before the assignment is completed.