[OTA-2061] Improve database lookup for list-installed-targets
The main problem here was that the EcuRepository.findall() method was performing a join
on the devices.primary_ecu_id which isn't indexed and therefore sloooooow. Updated it to join on
devices.id.

Also added a new function which joins all three tables: ecus, devices, and ecu_targets and use it when handling
list-installed-targets endpoint.

Signed-off-by: Ben Clouser <dev@benclouser.com>