Consider the correlationId provided in the installation report when sending messages to bus after receiving a device manifest
Previously, we relied on the reported installation report to match
current assignments and infer the correlation id. This behavior is
maintained, but in addition we extract the correlation id from the
manifest and send messages to the bus with that correlation id, if
it's different from matched correlation ids.

This change is required to support devices that install offline
updates and send a manifest when they eventually come online.

a libats bump is required to support the new lockbox/offline update
format for CorrelationId.