Add new endpoint to store an offline signed remote-sessions.json
`POST` on `:director/api/v1/admin/remote-sessions.json`

With an offline signed remote-sessions.json as body.

The remote-sessions.json body will be saved as received to the
database and provided to devices.

Director will validate the structure of remote-sessions.json and
verifiy the signatures of the payload.

The signed part will need to be signed with valid remote-sessions keys
defined in root.json. The user will need to have access to those
private keys. Currently, the only way to have access to these keys is
to rotate the root.json and add newly generated keys to the
namespace's new root.json. Read the `uptane-sign` documentation for
information on how to do this. Note that the existing remote-sessions
keys should not be removed from root.json, otherwise the user will
lose the ability to manage remote-sessions using the torizon web
UI.