check processed assignments when determining effects on scheduled updates
Previously, director used only the status of the ecus as reported by
the device to determine whether the scheduled update was
processed. However, cancelled assignments are not reflected on the ecu
status, so the processing of cancelled assignments originated from
scheduled updates was ignored.

These changes make the processed assignments be considered when
determining whether a scheduled update was processed or not.

Some `/debug` endpoints were fixed as there were no longer working.

Added a new main class `ManifestProcess` that can be called on the
CLI. This command receives a device state file and a manifest json
file and outputs the new state of the device, after the changes caused
by the given manifest are applied. This command does not have any
effects and works without a database. This means we can get a state
file from any environment, get a manifest file from the device, change
the manifest file, run the command, and see what changes would be
applied to the database. Since this works locally, we could change
the manifest compilation code and see what effect that would have on
the device state.

Two new endpoints added to `/debug` to debug the state of the device.

`/debug/device-state` dumps the state of the device in json
format. This state can the be used when running `ManifestProcess`.

`/debug/run-manifest` accepts a manifest and runs the compilation
process with the current state for the device, fetched from the
current database. This command has no effect on the database or the
device state, and simply outputs what would be the new state of the
device if the manifest was sent by the device at this moment.