Add search by tags
When searching by tags, the user can no longer use any other
filtering parameters.

The format of the tag values is restricted so it can be used with `=`
and prevent SQL injection, as we now need to dynamically build a query
string, since slick does not support searching by tuples, and
`(tag_id, tag_value)` is a tuple.

Adding more filtering methods is possible, but the dynamically build
query needs to be further complicated. The filtering methods to add
should be limited to the rows in the `Device` table.