Add `Seen` status to Update
Changes to `Seen` when its assignments move to `InFlight`. Note that
since we also check this status when cancelling an update, and this
check runs before the "assignments in flight" check, when an user
previously got a `assignment_in_flight_devices` error, they will now
get a `update_cannot_be_cancelled` error instead.

# Conflicts:
#	src/main/scala/com/advancedtelematic/director/db/UpdatesDBIO.scala